liblzma: Replace probability struct members in lzma decoder.
The probability struct members are replaced with a single contiguous
probability array. Many macros are defined to help index into this array
to find the proper sub-models.

The purpose of this is to try to improve speed. Using a single
contiguous array may give compilers a chance for additional
optimizations. In addition, this could be helpful in a future assembly
implementation since the assembly would not have to work with the struct
as much. The assembly version can easily copy the macros to find the
sub-models.

Many of the ideas for this were taken from Igor Pavlov's implementation
in LZMA SDK.