liblzma: Add overflow check for Unpadded size in lzma_index_append().
This was not a security bug since there was no path to overflow
UINT64_MAX in lzma_index_append() or when it calls index_file_size().
The bug was discovered by a failing assert() in index_file_size() when
unpadded_sum (the sum of the compressed size of previous streams and the
unpadded_size parameter) equals or exceeds LZMA_VLI_MAX.

Previously, the unpadded_size parameter was checked to be smaller than
UNPADDED_SIZE_MAX, but no check was done once compressed_base was added.

This could not have caused an integer overflow in index_file_size() when
called by lzma_index_append(). The calculation for file_size breaks down
into the sum of:

- Compressed base from all previous streams
- 2 * LZMA_STREAM_HEADER_SIZE (size of the current streams header and
  footer)
- stream_padding (can be set by lzma_index_stream_padding())
- Compressed base from the current stream
- Unpadded size (parameter to lzma_index_append())

The sum of everything except for Unpadded size must be less than
LZMA_VLI_MAX. This is guarenteed by overflow checks in the functions
that can set these values including lzma_index_stream_padding(),
lzma_index_append(), and lzma_index_cat(). The maximum value for
Unpadded size is enforced by lzma_index_append() to be less than or
equal UNPADDED_SIZE_MAX. Thus, the sum cannot exceed UINT64_MAX since
LZMA_VLI_MAX is half of UINT64_MAX.

Thanks to Joona Kannisto for reporting this.